Improvements across the board, see summary for details.
Weapons are now spawned in from prefabs and anchored to a specified position on the unit rather than being included as a child object and therefore inheriting the parent's rotation.

Target scanning now supports both random and nearest targeting modes. Random gives a more organic result and is faster due to the lack of a loop. Nearest may be useful for units such as snipers and point defenses but runs slower due to the need for a sorting loop.

Semi-custom selection shader implemented based on an online tutorial. Further work is needed to polish the look.

Randomization feature added to the Audio Manager to try and mitigate sounds piling up when audio source demand is high.

Imported Standard Assets - Prototyping to ain in testing and early development.